+++ CesiumAsync/include/CesiumAsync/SharedAssetDepot.h
@@ -47,7 +47,8 @@ public:
    *
    * Default is 16MiB.
    */
-  int64_t inactiveAssetSizeLimitBytes = static_cast<int64_t>(16 * 1024 * 1024);
+  std::atomic<int64_t> inactiveAssetSizeLimitBytes =
+      static_cast<int64_t>(16 * 1024 * 1024);
 
   /**
    * @brief Signature for the callback function that will be called to fetch and
@@ -94,6 +95,18 @@ public:
       const AsyncSystem& asyncSystem,
       const std::shared_ptr<IAssetAccessor>& pAssetAccessor,
       const TAssetKey& assetKey);
+
+  /**
+   * @brief Invalidates the previously-cached asset with the given key, so that
+   * the next call to {@link getOrCreate} will create the asset instead of
+   * returning the existing one.
+   *
+   * Anyone already using the existing asset may continue to do so.
+   *
+   * If an asset with the given key does not exist in the depot, this method
+   * does nothing.
+   */
+  void invalidate(const TAssetKey& assetKey);
 
   /**
    * @brief Returns the total number of distinct assets contained in this depot,
@@ -254,6 +267,11 @@ private:
   // list.
   int64_t _totalDeletionCandidateMemoryUsage;
 
+  // The number of assets that have been invalidated but that have not been
+  // deleted yet. Such assets hold a pointer to the depot, so the depot must be
+  // kept alive for their entire lifetime.
+  int64_t _liveInvalidatedAssets;
+
   // Mutex serializing access to _assets, _assetsByPointer, _deletionCandidates,
   // and any AssetEntry owned by this depot.
   mutable std::mutex _mutex;
@@ -275,6 +293,7 @@ SharedAssetDepot<TAssetType, TAssetKey>::SharedAssetDepot(
       _assetsByPointer(),
       _deletionCandidates(),
       _totalDeletionCandidateMemoryUsage(0),
+      _liveInvalidatedAssets(0),
       _mutex(),
       _factory(std::move(factory)),
       _pKeepAlive(nullptr) {}
@@ -302,6 +321,7 @@ SharedAssetDepot<TAssetType, TAssetKey>::~SharedAssetDepot() {
   // this destructor from being called except when all of its assets are also
   // in the _deletionCandidates list.
 
+  CESIUM_ASSERT(this->_liveInvalidatedAssets == 0);
   CESIUM_ASSERT(this->_assets.size() == this->_deletionCandidates.size());
 }
 
@@ -402,6 +422,46 @@ SharedAssetDepot<TAssetType, TAssetKey>::getOrCreate(
 }
 
 template <typename TAssetType, typename TAssetKey>
+void SharedAssetDepot<TAssetType, TAssetKey>::invalidate(
+    const TAssetKey& assetKey) {
+  LockHolder lock = this->lock();
+
+  auto it = this->_assets.find(assetKey);
+  if (it == this->_assets.end())
+    return;
+
+  AssetEntry* pEntry = it->second.get();
+  CESIUM_ASSERT(pEntry);
+
+  // This will remove the asset from the deletion candidates list, if it's
+  // there.
+  CesiumUtility::ResultPointer<TAssetType> assetResult =
+      pEntry->toResultUnderLock();
+
+  if (assetResult.pValue) {
+    if (!assetResult.pValue->_isInvalidated) {
+      assetResult.pValue->_isInvalidated = true;
+      ++this->_liveInvalidatedAssets;
+    }
+    this->_assetsByPointer.erase(assetResult.pValue.get());
+  }
+
+  // Detach the asset from the AssetEntry, so that its lifetime is controlled by
+  // reference counting.
+  pEntry->pAsset.release();
+
+  // Remove the asset entry. This won't immediately delete the asset, because
+  // `assetResult` above still holds a reference to it. But once that goes out
+  // of scope, too, the asset _may_ be destroyed.
+  this->_assets.erase(it);
+
+  // Unlock the mutex before allowing `assetResult` to go out of scope. When it
+  // goes out of scope, the asset may be destroyed. If it is, that would cause
+  // us to try to re-enter the lock, which is not allowed.
+  lock.unlock();
+}
+
+template <typename TAssetType, typename TAssetKey>
 size_t SharedAssetDepot<TAssetType, TAssetKey>::getAssetCount() const {
   LockHolder lock = this->lock();
   return this->_assets.size();
@@ -448,6 +508,21 @@ void SharedAssetDepot<TAssetType, TAssetKey>::markDeletionCandidate(
 template <typename TAssetType, typename TAssetKey>
 void SharedAssetDepot<TAssetType, TAssetKey>::markDeletionCandidateUnderLock(
     const TAssetType& asset) {
+  if (asset._isInvalidated) {
+    // This asset is no longer tracked by the depot, so delete it.
+    --this->_liveInvalidatedAssets;
+    delete &asset;
+
+    // If this depot is not managing any live assets, then we no longer need to
+    // keep it alive.
+    if (this->_assets.size() == this->_deletionCandidates.size() &&
+        this->_liveInvalidatedAssets == 0) {
+      this->_pKeepAlive.reset();
+    }
+
+    return;
+  }
+
   // Verify that the reference count is still zero.
   // See: https://github.com/CesiumGS/cesium-native/issues/1073
   if (asset._referenceCount != 0) {
@@ -494,7 +569,8 @@ void SharedAssetDepot<TAssetType, TAssetKey>::markDeletionCandidateUnderLock(
 
   // If this depot is not managing any live assets, then we no longer need to
   // keep it alive.
-  if (this->_assets.size() == this->_deletionCandidates.size()) {
+  if (this->_assets.size() == this->_deletionCandidates.size() &&
+      this->_liveInvalidatedAssets == 0) {
     this->_pKeepAlive.reset();
   }
 }
@@ -514,6 +590,12 @@ void SharedAssetDepot<TAssetType, TAssetKey>::unmarkDeletionCandidate(
 template <typename TAssetType, typename TAssetKey>
 void SharedAssetDepot<TAssetType, TAssetKey>::unmarkDeletionCandidateUnderLock(
     const TAssetType& asset) {
+  // This asset better not already be invalidated. That would imply this asset
+  // was resurrected after its reference count hit zero. This should only be
+  // possible if the asset depot returned a pointer to the asset, which it
+  // will not do for one that is invalidated.
+  CESIUM_ASSERT(!asset._isInvalidated);
+
   auto it = this->_assetsByPointer.find(const_cast<TAssetType*>(&asset));
   CESIUM_ASSERT(it != this->_assetsByPointer.end());
   if (it == this->_assetsByPointer.end()) {